+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## 12/15/2021 2.11.4
 
+- LiveValidator - Validate multipleOf data again when ajv validation results have error about multipleOf in response
 - LiveValidator - Ignore required validation rule in response when x-ms-secret is true and the httpMethod is POST
 - ModelValidator - add validation for common path scope parameters in swagger
 - ModelValidator - Increase model validation test coverage to 90%+